requirements.txt
flask
flask_sqlalchemy
flask_login
flask_dance
python-dotenv
requests
gunicorn
google-generativeai
